fix: Don't return default sink in streams getter
This prevents weird timing issues when muting/unmuting and the Spotify streams have already closed or haven't opened yet. Also check for Spotify streams specifically in mute/unmute logic and wrap all the logic in the check, again to prevent any kind of timing issue where the Spotify stream is opened or closed in the middle of the logic.